# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TripLog API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the user profile",
                "description": "Allows updating the user's profile including email, name, avatar, bio, and social media links.",
                "operationId": "ede45b19680d6fb1319013a741a05734",
                "responses": {
                    "200": {
                        "description": "Successful profile retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://example.com/avatar.jpg"
                                        },
                                        "facebook_link": {
                                            "type": "string",
                                            "example": "https://facebook.com/johndoe"
                                        },
                                        "instagram_link": {
                                            "type": "string",
                                            "example": "https://instagram.com/johndoe"
                                        },
                                        "x_link": {
                                            "type": "string",
                                            "example": "https://x.com/johndoe"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "Traveler and photographer"
                                        },
                                        "travel_preferences": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "mountains",
                                                "beaches"
                                            ]
                                        },
                                        "trips_count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "planned_trips_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the user profile",
                "description": "Allows updating the user's profile including email, name, avatar, bio, and social media links.",
                "operationId": "b37e1c285c5fcd621538c248c3f414a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "avatar": {
                                        "description": "Avatar image file (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "facebook_link": {
                                        "type": "string",
                                        "example": "https://facebook.com/johndoe"
                                    },
                                    "instagram_link": {
                                        "type": "string",
                                        "example": "https://instagram.com/johndoe"
                                    },
                                    "x_link": {
                                        "type": "string",
                                        "example": "https://x.com/johndoe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Traveler and photographer"
                                    },
                                    "travel_preferences": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "mountains",
                                            "beaches",
                                            "forests"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://example.com/avatar.jpg"
                                        },
                                        "facebook_link": {
                                            "type": "string",
                                            "example": "https://facebook.com/johndoe"
                                        },
                                        "instagram_link": {
                                            "type": "string",
                                            "example": "https://instagram.com/johndoe"
                                        },
                                        "x_link": {
                                            "type": "string",
                                            "example": "https://x.com/johndoe"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "Traveler and photographer"
                                        },
                                        "travel_preferences": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "mountains",
                                                "beaches"
                                            ]
                                        },
                                        "trips_count": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "planned_trips_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name must not exceed 32 characters."
                                                    }
                                                },
                                                "avatar": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The avatar must be an image."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer $token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}